Fix potential subtraction overflow
When doing test runs the interval snapshotting
resulted in a panic because of unsigned integer underflow.
Apparently in the traceset data, at least one of
- read bytes
- write bytes
- blkiodelay
did decrease from one snapshot to the other.

This should only happen when a newer snapshot has
fewer/different tracees, but then it should be invalid.
Somehow it happened for a valid snapshot, so now
the subtraction method is safe and will return 0 if the
second snapshot's value is lower than the previous.